src/identity-service/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
fastapi
uvicorn
databases
sqlalchemy
asyncpg
